Fixed typo that breaks the build. G_GNUC_CONST'ified *_get_type(); Marked
	* src/cpufreq-applet.c: (cpufreq_applet_about_cb):
	* src/cpufreq-applet.h:
	* src/cpufreq-monitor-cpuinfo.c: (cpufreq_monitor_cpuinfo_get):
	* src/cpufreq-monitor-cpuinfo.h:
	* src/cpufreq-monitor-procfs.c: (cpufreq_monitor_procfs_parse):
	* src/cpufreq-monitor-procfs.h:
	* src/cpufreq-monitor-sysfs.h:
	* src/cpufreq-monitor.h:
	* src/cpufreq-prefs.c: (cpufreq_prefs_cpu_combo_setup):
	* src/cpufreq-selector/cpufreq-procfs.h:
	* src/cpufreq-selector/cpufreq-sysfs.c:
	* src/cpufreq-selector/cpufreq-sysfs.h:
	* src/cpufreq-selector/cpufreq.h:
	* src/cpufreq-selector/main.c: (main):

	Fixed typo that breaks the build.
	G_GNUC_CONST'ified *_get_type();
	Marked const some strings.
	Fixed signedness warnings.
	Added missing #include.

+++ cpufreq/src/cpufreq-monitor-cpuinfo.c
@@ -160,7 +160,7 @@ cpufreq_monitor_cpuinfo_get (gpointer gdata)
                                 return FALSE;
                         }
 
-                        if (strlen (lines[i]) < (p - lines[i])) {
+                        if (strlen (lines[i]) < (size_t)(p - lines[i])) {
                                 g_strfreev (lines);
                                 g_free (file);
